refactor: modularize bot and add tests
